Tested codex original theme code for social icons
I took all the example code
from the codex theme repo and pasted it
into my own repo for my website to see
whether or not the social icons would appear
and they did not. So I am not sure what the
hell the problem is and how to resolve it.
Reverting back to the latest version of my
own site.