Figured out all the problems
Things weren't loading right, the home page was appearing as
the projects page, then the projects page wasn't appearing at all.
Last time I made this webiste the issue was theat the social icons
on the home page and at the bottom of each page were not loading.
this time I got them to work on the first try but I forgot to add
the new images files and the new page directory to the repo.
They were not in the tracked files and so when I did "git add -u"
they didn't get added in and so hugo didn't know what was up.
Remember next time, for the social icons go into "about/index.html"
and change all the https urls to http. Then make sure to add in any
new files to your commit that you want to load on the site. That
means both pages (in the form of new subdirectories) and content
such as images. The website is looking good. Love the russian blue.